+++ examples/next/pages/pages/index.tsx
@@ -1,10 +1,16 @@
 import type { GetServerSidePropsContext, NextPage } from 'next'
+import { RangeSlider, Slider } from '@yamada-ui/react'
 import { Layout } from 'components/layouts'
 
 interface Props {}
 
 const Page: NextPage<Props> = () => {
-  return <Layout />
+  return (
+    <Layout>
+      <RangeSlider />
+      <Slider />
+    </Layout>
+  )
 }
 
 export default Page