+++ packages/components/calendar/src/date-picker.tsx
@@ -2,7 +2,7 @@ import type { CSSUIObject, FC, HTMLUIProps, ThemeProps } from "@yamada-ui/core"
 import type { IconProps } from "@yamada-ui/icon"
 import type { MotionProps } from "@yamada-ui/motion"
 import type { PortalProps } from "@yamada-ui/portal"
-import type { HTMLAttributes, ReactNode, RefAttributes } from "react"
+import type { ReactNode, RefAttributes } from "react"
 import type { UseDatePickerProps } from "./use-date-picker"
 import {
   forwardRef,
@@ -187,7 +187,7 @@ DatePicker.displayName = "DatePicker"
 DatePicker.__ui__ = "DatePicker"
 
 interface DatePickerFieldOptions {
-  inputProps?: HTMLUIProps<"input">
+  inputProps?: HTMLUIProps<"input"> & RefAttributes<HTMLInputElement>
 }
 
 export interface DatePickerFieldProps
@@ -197,8 +197,7 @@ export interface DatePickerFieldProps
 export const DatePickerField = forwardRef<DatePickerFieldProps, "input">(
   ({ className, h, minH, inputProps, ...rest }, ref) => {
     const styles = useDatePickerContext()
-    const { ref: inputRef, ...computedInputProps } =
-      inputProps as HTMLUIProps<"input"> & RefAttributes<HTMLInputElement>
+    const { ref: inputRef, ...computedInputProps } = inputProps ?? {}
 
     const css: CSSUIObject = {
       alignItems: "center",
@@ -322,7 +321,7 @@ export const DatePickerClearIcon: FC<DatePickerClearIconProps> = ({
   const rest = useClickable({
     ref,
     isDisabled,
-    ...(props as HTMLAttributes<HTMLElement>),
+    ...props,
   })
 
   return (