+++ packages/components/fontawesome/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @yamada-ui/fontawesome
 
+## 1.0.42
+
+### Patch Changes
+
+- [#3012](https://github.com/yamada-ui/yamada-ui/pull/3012) [`3826e1f`](https://github.com/yamada-ui/yamada-ui/commit/3826e1fa058ed091d149a177ca4aaccadf8d833c) Thanks [@Tetsu-is](https://github.com/Tetsu-is)! - add `displayName` to Icon Component
+
+- Updated dependencies [[`898ab83`](https://github.com/yamada-ui/yamada-ui/commit/898ab83eb677646ee293a336a0a5d254aeefa00f), [`2173466`](https://github.com/yamada-ui/yamada-ui/commit/2173466c144212522a61aec94e34cac375291655), [`28632a8`](https://github.com/yamada-ui/yamada-ui/commit/28632a8d904b581068e93a75e44323c9d37111b1), [`c22d335`](https://github.com/yamada-ui/yamada-ui/commit/c22d3355ab42d4f0f3e3413248bed949729fc8e3), [`4832b3d`](https://github.com/yamada-ui/yamada-ui/commit/4832b3d8d6184fb36c7f06fed2c8ef51e4d6542b)]:
+  - @yamada-ui/core@1.15.2
+  - @yamada-ui/utils@1.5.3
+  - @yamada-ui/icon@1.1.8
+
 ## 1.0.41
 
 ### Patch Changes